[Minor] Cleanup and evaluate checks in batches
